+++ popkorn-androidx-viewmodel/src/androidTest/kotlin/data/TestViewModel.kt
@@ -1,0 +1,5 @@
+package cc.popkorn.androidx.viewModel.data
+
+import androidx.lifecycle.ViewModel
+
+class TestViewModel : ViewModel()